+++ napari/_vispy/experimental/vispy_image_layer2.py
@@ -156,9 +156,9 @@ class VispyImageLayer2(VispyBaseLayer):
 
         self.node.visible = False if self.layer._empty else self.layer.visible
 
-        self._on_scale_change()
-        self._on_translate_change()
-        node.update()
+        # Call to update order of translation values with new dims:
+        self._on_matrix_change()
+        self.node.update()
 
     def _on_interpolation_change(self, event=None):
         self.node.interpolation = self.layer.interpolation
@@ -231,7 +231,7 @@ class VispyImageLayer2(VispyBaseLayer):
             for i, d in enumerate(self.layer.dims.displayed):
                 scale[d] = downsample[i]
             self.layer._transforms['tile2data'].scale = scale
-            self._on_scale_change()
+            self._on_matrix_change()
             slices = tuple(slice(None, None, ds) for ds in downsample)
             data = data[slices]
         return data
